+++ core/src/test/java/arez/SpyImplTest.java
@@ -420,23 +420,6 @@ public class SpyImplTest
   }
 
   @Test
-  public void observable_hasAccessor_introspectorsDisabled()
-    throws Throwable
-  {
-    ArezTestUtil.disablePropertyIntrospectors();
-
-    final ArezContext context = Arez.context();
-    final Spy spy = context.getSpy();
-
-    final ObservableValue<Integer> observableValue = context.observable();
-
-    final IllegalStateException exception2 =
-      expectThrows( IllegalStateException.class, () -> context.action( () -> spy.hasAccessor( observableValue ) ) );
-    assertEquals( exception2.getMessage(),
-                  "Arez-0110: Spy.hasAccessor invoked when Arez.arePropertyIntrospectorsEnabled() returns false." );
-  }
-
-  @Test
   public void observable_hasMutator_introspectorsDisabled()
     throws Throwable
   {